Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
asyncpg
python-dotenv
supabase
python-multipart
